RFC: Modified agenda layout
This is a concept at how we can approach the agenda with a hopefully
clearer approach and better support for mobile layout. It uses Grid
so we are losing IE 11 but there are bigger issues to resolve there. It
wouldn't be overly difficult to support this layout with Flex when/if
required