RFC: Adds session feedback page
Would like feedback on the approach I have taken for this. Things that I
have introduced:

useSession hook
handles front end loading of Sessions. Added to Agenda and Feedback page
Mostly similar to what we have no but in hook form with reducer. Also
houses the loading code for SSR. I feel like this could be joined together
somehow because they're repeating. Any input would be good. Suspect CSR
supports async so it might basically be there

useSessionGroup hook
This is probably the biggest thing I need eyes on. The purpose is to keep
the sessions in groups and then be able to know what the past sessions,
previous sessions, current sessions, next sessions and future sessions are.
The gaps made this more difficult than I would have liked

How I see this being used:
* Show the previous sessions on the feedback page (done)
* Past sessions in a show/hide under the previous session on feedback page (done)
* Show the next sessions at the top of the agenda page without needing to hunt for it (todo)

The code for calculating it seems like it could use improvement. I have
tried several variations on this and trying to memoize it because of how
frequently it runs. The current version is great for memoization but maybe
a little gross. Setting up groups also leaves a little to be desired.

useForm hook
Hook that handles changing in form and calling a submit callback.

storageKey helper
Small helper function when we are using local storage

Finally is the feedback page. I have added test buttons here to be able to
jump between times to see the sessions. I have gone with the traffic like
approach but need input on how to display that.

TL;DR
* Hooks all the things
* Can I improve useSessionGroups? Particularly the logic calculating the state
* Feedback page - what do we think?