Fixes #217 intermittent agenda modal styling issue
I had added the ReachUI styles to next.js config and imported the css but
that appears to have had issues at certain points. After investigating
what was actually in that stylesheet we were using very little of it after
having otherwise overwritten the properties. The properties that we were
using are now in Emotion and the changes to next's config reverted.