Chore: replace deprecated TSLint with ESLint
TSLint is deprecated (https://medium.com/palantir/tslint-in-2019-1a144c2317a9)
in favour of ESLint. This PR changes the lint to ESLint using the open
source eslint-config-wanews as the base.

This has meant there are variations in the rules but I feel that they're
a little more useful than before (e.g. removal of the object order rule)
Rules are descriped in `.eslintrc`.

* Replaces TSLint with ESLint
* Uses eslint-config-wanews as base
* Adds react and react-hook rules
* Removes TSLint disable comments
* Adds disable rules where there's confusing (needs review)