Resolving build issues
The CI pipeline isn't working due to node-gyp errors. Cause appears to
be related to node-sass and the version of Node that the CI is using.
Recreated the issue locally by updating node to v12. The only way I
could get it to succeed was nuking the node_modules and yarn lock and
starting over.

That means there's a lot of verion changes. The site appears to be ok
but have a look and poke around to see if anything isn't working.