Perf: Replace Moment with date-fns
Moment.js is a weighty package considering we use a small fraction of it
Date-fns is a better option, especially given its smaller size. PR
replaces Moment.js with the equivalent date-fns method.

Deletes the eventDetails, countdownTimer and countdown component as we
aren't using them. They were throwing various React errors so if we
reimplement this feature we can create a more modern component.