Remove withPageMetadata and withCurrentDate
The two Higher Order Components made it impossible to use newer features
of Next and React. To resolve this, this PR restructures what they
provided into the Config hook. Within the hook is the configuration and
timing.

This opened up the possibility of using GetServerSideProps and all
functionality that could be moved there was. This will mean most data
loading is handled on the server along with handling errors and the
page being unavailable. Any code that's used only on the server will
then be omitted from the front end bundle.

This will allow easier integration with Feature Toggles and starting to
move pages that don't need to be rendered client side to static pages.
Previous agendas, about and so on.

There's a lot of rewiring here. Testing would be great.