Privacy, vote landing and rewrites
We needed a vote landing page, a privacy page and to default to the
elo voting system instead of the older system.

Creates the privacy policy page, using the content in Teams.

Creates the vote landing page which will write a cookie that signifies
user's acceptance of the privacy policy. A check was added to the elo
voting page to redirect users to the landing page if the cookie isn't
present.

Rewrite configuration was created to keep the link structure the same
but use the elo landing and voting system instead of the existing system